requirements.txt
pandas
boto3
google-cloud
apache-airflow[amazon]
apache-airflow[google]
